backend/requirements.txt
pageindex
docling
python-dotenv
google-genai
langgraph
fastapi
uvicorn[standard]
python-multipart
torch
transformers
groq
